Rework `Cooked.Pretty` module structure
